Splitting code depending on dataset.
