+++ experiment_SHD/train.py
@@ -1,14 +1,14 @@
-from simulator_SHD import *
+from simulator import *
 import matplotlib.pyplot as plt
 import numpy as np
 import json
 
-p["TRIAL_MS"]= 1400
+p["TRIAL_MS"]= 400
 p["DATASET"]= "SHD"
 p["TRAIN_DATA_SEED"]= 372
 p["TEST_DATA_SEED"]= 814
 p["MODEL_SEED"]= 135
-p["NAME"]= "test101"
+p["NAME"]= "test43"
 p["NUM_HIDDEN"]= 256
 p["N_MAX_SPIKE"]= 1500
 p["DT_MS"]= 1
@@ -19,12 +19,12 @@ p["ADAM_BETA2"]= 0.999
 p["DEBUG"]= False
 p["DEBUG_HIDDEN_N"]= True
 p["LOAD_LAST"]= False
-p["N_EPOCH"]= 200
-p["N_BATCH"]= 256
+p["N_EPOCH"]= 10
+p["N_BATCH"]= 32
 p["SUPER_BATCH"]= 1
 p["N_TRAIN"]= 7900 #20*p["N_BATCH"] #7756 
 p["N_VALIDATE"]= 512 # 256 # p["N_BATCH"] 
-p["ETA"]= 5e-3 #1e-4 # 5e-3
+p["ETA"]= 2e-3 #1e-4 # 5e-3
 p["SHUFFLE"]= True
 p["INPUT_HIDDEN_MEAN"]= 0.02 # 0.02
 p["INPUT_HIDDEN_STD"]= 0.01 # 0.01
@@ -34,9 +34,9 @@ p["W_REPORT_INTERVAL"] = 11000  # this should be at the end of the epoch (at fir
 p["TAU_MEM"] = 20.0 #20
 p["TAU_SYN"] = 5.0 #5
 p["REG_TYPE"]= "simple"
-p["LBD_UPPER"]= 2e-14 # 2e-9 # 2e-8 # 2e-14 (since removal of N_Batch), 5e-12 keep in mind that the term is applied to all contributing spikes ...
-p["LBD_LOWER"]= 2e-14 #2e-8
-p["NU_UPPER"]= 14
+p["LBD_UPPER"]= 0 # 2e-9 # 2e-8 # 2e-14 (since removal of N_Batch), 5e-12 keep in mind that the term is applied to all contributing spikes ...
+p["LBD_LOWER"]= 2e-8 #2e-8
+p["NU_UPPER"]= 15 #*p["N_BATCH"]
 p["NU_LOWER"]= 5
 p["RHO_UPPER"]= 10000.0
 p["GLB_UPPER"]= 1e-8
@@ -48,22 +48,22 @@ p["TIMING"]= True
 p["SPK_REC_STEPS"]= int(p["TRIAL_MS"]/p["DT_MS"])
 p["EVALUATION"]= "speaker"
 
-p["RECURRENT"]= True
-p["HIDDEN_HIDDEN_MEAN"]= 0.0
-p["HIDDEN_HIDDEN_STD"]= 0.02 # 0.02
-
-p["REWIRE_SILENT"]= True
-p["AVG_SNSUM"]= True
-
 # "first_spike" loss function variables
 p["TAU_0"]= 0.5
 p["TAU_1"]= 6.4
 p["ALPHA"]= 3e-3
 
-#p["LOSS_TYPE"]= "first_spike"
+p["LOSS_TYPE"]= "first_spike" # TODO: in progress...
 #p["LOSS_TYPE"]= "max"
-p["LOSS_TYPE"]= "sum"
+#p["LOSS_TYPE"]= "sum"
 #p["LOSS_TYPE"]= "avg_xentropy"
+
+p["RECURRENT"]= False
+p["HIDDEN_HIDDEN_MEAN"]= 0.0
+p["HIDDEN_HIDDEN_STD"]= 0.02 # 0.02
+
+p["REWIRE_SILENT"]= False
+p["AVG_SNSUM"]= True
 
 #p["AUGMENTATION"]= {}
 
@@ -75,7 +75,7 @@ p["AUGMENTATION"]= {
 
 #p["REDUCED_CLASSES"]= [0]
 
-"""
+
 p["REC_NEURONS"]= [("output","V"),("output","sum_V"),("output","lambda_V"),("output","lambda_I"),
                    ("hidden","lambda_V")]
 p["REC_NEURONS_EPOCH_TRIAL"]= [(0,25),(0,26),(0,27),(0,28),(0,29),(0,30),(0,31),
@@ -90,7 +90,6 @@ p["REC_SPIKES"]= ["input","hidden"]
 p["W_OUTPUT_EPOCH_TRIAL"]= [(0,25),(0,26),(0,27),(0,28),(0,29),(0,30),(0,31),
                               (9,25),(9,26),(9,27),(9,28),(9,29),(9,30),(9,31),
 ]
-"""
 
 if p["DEBUG"]:
     p["REC_SPIKES"]= ["input", "hidden"]
@@ -112,7 +111,7 @@ p["W_OUTPUT_EPOCH_TRIAL"]= [(0,0), (100,0), (799,0)]
 with open(os.path.join(p["OUT_DIR"], p["NAME"]+'.json'), 'w') as file:
     json.dump(p, file)
     
-mn= SHD_model(p)
+mn= mnist_model(p)
 spike_t, spike_ID, rec_vars_n, rec_vars_s,correct,correct_eval= mn.train(p)
 
 print("correct: {}".format(correct))
